+++ types/interfaces.d.ts
@@ -5,7 +5,7 @@ export interface ISOParts {
   M: number;
   /** Week designator that follows the value for the number of weeks */
   W: number;
-  /** Day designator that follows the value for the number of calendar days */
+  /** Day designator that follows the value for the number of calendar days, or weekday if week designator is used */
   D: number;
   /** Hour designator that follows the value for the number of hours */
   H: number;
@@ -26,6 +26,25 @@ export interface ISODateParts extends ISOParts {
   Om?: number;
   /** Seconds offset as of ISO 8601-2:2019 */
   OS?: number;
+}
+
+export enum ISOWeekday {
+  Monday = 1,
+  Tuesday = 2,
+  Wednesday = 3,
+  Thursday = 4,
+  Friday = 5,
+  Saturday = 6,
+  Sunday = 7,
+}
+
+export interface ISOWeekParts {
+  /** Year */
+  Y: number;
+  /** Weeknumber */
+  W: number;
+  /** ISO weekday */
+  weekday: ISOWeekday;
 }
 
 export interface ISOInterval extends ISOParts {